Move ZSeamConfig to separate file
This is required by our code style.

Contributes to issue CURA-7558.
