+++ src/PathOrderOptimizer.h
@@ -5,63 +5,11 @@
 #define PATHORDEROPTIMIZER_H
 
 #include "settings/EnumSettings.h" //To get the seam settings.
+#include "settings/ZSeamConfig.h" //To read the seam configuration.
 #include "utils/polygonUtils.h"
 #include "utils/linearAlg2D.h" //To find the angle of corners to hide seams.
 
 namespace cura {
-
-/*!
- * Helper class that encapsulates the various criteria that define the location
- * of the z-seam.
- * Instances of this are passed to the PathOrderOptimizer to specify where the
- * seam is to be located.
- */
-struct ZSeamConfig
-{
-    /*!
-     * Strategy to place the seam (user-specified, shortest distance, sharpest
-     * corner, etc.).
-     */
-    EZSeamType type;
-
-    /*!
-     * When using a user-specified position for the seam, this is the position
-     * that the user specified.
-     */
-    Point pos;
-
-    /*!
-     * Corner preference type, if using the sharpest corner strategy.
-     */
-    EZSeamCornerPrefType corner_pref;
-
-    /*!
-     * Default constructor for use when memory must be allocated before it gets
-     * filled (like with some data structures).
-     *
-     * This will select the "shortest" seam strategy.
-     */
-    ZSeamConfig()
-    : type(EZSeamType::SHORTEST)
-    , pos(Point(0, 0))
-    , corner_pref(EZSeamCornerPrefType::Z_SEAM_CORNER_PREF_NONE)
-    {
-    }
-
-    /*!
-     * Create a seam configuration with a custom configuration.
-     * \param type The strategy to place the seam.
-     * \param pos The position of a user-specified seam.
-     * \param corner_pref The corner preference, when using the sharpest corner
-     * strategy.
-     */
-    ZSeamConfig(const EZSeamType type, const Point pos, const EZSeamCornerPrefType corner_pref)
-    : type(type)
-    , pos(pos)
-    , corner_pref(corner_pref)
-    {
-    }
-};
 
 /*!
  * Path order optimization class.
