+++ src/settings/ZSeamConfig.h
@@ -1,0 +1,58 @@
+//Copyright (c) 2020 Ultimaker B.V.
+//CuraEngine is released under the terms of the AGPLv3 or higher.
+
+#ifndef ZSEAMCONFIG_H
+#define ZSEAMCONFIG_H
+
+#include "EnumSettings.h" //For EZSeamType and EZSeamCornerPrefType.
+#include "../utils/IntPoint.h" //To store the preferred seam position.
+
+namespace cura
+{
+
+/*!
+ * Helper class that encapsulates the various criteria that define the location
+ * of the z-seam.
+ * Instances of this are passed to the PathOrderOptimizer to specify where the
+ * seam is to be located.
+ */
+struct ZSeamConfig
+{
+    /*!
+     * Strategy to place the seam (user-specified, shortest distance, sharpest
+     * corner, etc.).
+     */
+    EZSeamType type;
+
+    /*!
+     * When using a user-specified position for the seam, this is the position
+     * that the user specified.
+     */
+    Point pos;
+
+    /*!
+     * Corner preference type, if using the sharpest corner strategy.
+     */
+    EZSeamCornerPrefType corner_pref;
+
+    /*!
+     * Default constructor for use when memory must be allocated before it gets
+     * filled (like with some data structures).
+     *
+     * This will select the "shortest" seam strategy.
+     */
+    ZSeamConfig();
+
+    /*!
+     * Create a seam configuration with a custom configuration.
+     * \param type The strategy to place the seam.
+     * \param pos The position of a user-specified seam.
+     * \param corner_pref The corner preference, when using the sharpest corner
+     * strategy.
+     */
+    ZSeamConfig(const EZSeamType type, const Point pos, const EZSeamCornerPrefType corner_pref);
+};
+
+} //Cura namespace.
+
+#endif //ZSEAMCONFIG_H
